Fixed Player Animation not Updating Properly when Changing Floors

+++ Transluminal/Assets/Scripts/Game Controller/FoodManager.cs
@@ -90,7 +90,7 @@ public class FoodManager : MonoBehaviour
         return currentFood;
     }
 
-    public void SubtractFood(int food)
+    public void SubtractFood(float food)
     {
         currentFood -= food;
     }
